move sample_weights to skill_checks
