+++ src/models/scenes/scenes_base.py
@@ -1,9 +1,8 @@
 """Basic interfaces for scenes, effects, and resolutions."""
 import abc
-import random
-from bisect import bisect_left
-from itertools import accumulate
-from typing import Callable, Sequence, Tuple, Any
+from typing import Callable, Sequence, Tuple
+
+from models.scenes.skill_checks import sample_weights
 
 
 class Scene(metaclass=abc.ABCMeta):
@@ -51,33 +50,6 @@ class BasicResolution(Resolution):
     @property
     def effects(self) -> Sequence[EffectType]:
         return self._effect_seq
-
-
-def sample_weights(weighted_objects: Sequence[Tuple[Any, int]]) -> Any:
-    """Return a probabilistic sample based on weights.
-
-    Each possible outcome is given a positive weight. Its relative
-    probability is its weight divided by the sum of all weights.
-
-    Args:
-        weighted_objects: A sequence of outcome,weight pairs.
-    Returns:
-        One of the outcomes in weighted_objects, with probability matching its
-        relative weight.
-    """
-
-    assert weighted_objects, 'At least one resolution must be specified.'
-    assert all(rw[1] >= 0 for rw in weighted_objects), (
-        'weights must be positive, got {}'.format(weighted_objects))
-
-    # Sample outcomes according to weight.
-    cum_weights = list(accumulate((rw[1] for rw in weighted_objects),
-                                  lambda a, b: a + b))
-    num = random.randint(0, cum_weights[-1] - 1) + 1
-    index = bisect_left(cum_weights, num)
-
-    assert index < len(weighted_objects), (num, index)
-    return weighted_objects[index][0]
 
 
 class ProbabilisticResolution(Resolution):
