+++ src/models/scenes/skill_checks.py
@@ -1,11 +1,12 @@
 """Implementation of skill checks."""
 import random
+from bisect import bisect_left
 from enum import Enum
-from typing import Sequence, Union
+from itertools import accumulate
+from typing import Sequence, Tuple, Any
 
 from models.characters.player import get_player
-from models.characters.states import Skills, AttributeType
-from models.scenes.scenes_base import Scene, SceneConstructor
+from models.characters.states import AttributeType
 
 
 class Difficulty(Enum):
@@ -39,3 +40,30 @@ class Difficulty(Enum):
 
 _difficulty_probs = {k: v for k, v in
                      zip(Difficulty, [0, 1 / 8, 1 / 4, 1 / 2, 3 / 4, 7 / 8, 1])}
+
+
+def sample_weights(weighted_objects: Sequence[Tuple[Any, int]]) -> Any:
+    """Return a probabilistic sample based on weights.
+
+    Each possible outcome is given a positive weight. Its relative
+    probability is its weight divided by the sum of all weights.
+
+    Args:
+        weighted_objects: A sequence of outcome,weight pairs.
+    Returns:
+        One of the outcomes in weighted_objects, with probability matching its
+        relative weight.
+    """
+
+    assert weighted_objects, 'At least one resolution must be specified.'
+    assert all(rw[1] >= 0 for rw in weighted_objects), (
+        'weights must be positive, got {}'.format(weighted_objects))
+
+    # Sample outcomes according to weight.
+    cum_weights = list(accumulate((rw[1] for rw in weighted_objects),
+                                  lambda a, b: a + b))
+    num = random.randint(0, cum_weights[-1] - 1) + 1
+    index = bisect_left(cum_weights, num)
+
+    assert index < len(weighted_objects), (num, index)
+    return weighted_objects[index][0]
